+++ src/js/setup-text-tracks.js
@@ -1,8 +1,9 @@
 import dashjs from 'dashjs';
 import videojs from 'video.js';
+import window from 'global/window';
 
 function find(l, f) {
-  for(let i = 0; i < l.length; i++) {
+  for (let i = 0; i < l.length; i++) {
     if (f(l[i])) {
       return l[i];
     }
@@ -30,12 +31,14 @@ function attachDashTextTracksToVideojs(player, tech, tracks) {
         label: track.lang,
         language: track.lang,
         srclang: track.lang,
-      },
+        kind: track.kind
+      }
     }))
 
     // Add track to videojs track list
     .map(({trackConfig, dashTrack}) => {
-      const remoteTextTrack = player.addRemoteTextTrack(trackConfig, true);
+      const remoteTextTrack = player.addRemoteTextTrack(trackConfig, false);
+
       trackDictionary.push({textTrack: remoteTextTrack.track, dashTrack});
 
       // Don't add the cues becuase we're going to let dash handle it natively. This will ensure
@@ -46,6 +49,7 @@ function attachDashTextTracksToVideojs(player, tech, tracks) {
 
       return remoteTextTrack;
     })
+
   ;
 
   /*
@@ -90,7 +94,7 @@ function attachDashTextTracksToVideojs(player, tech, tracks) {
   player.textTracks().on('change', updateActiveDashTextTrack);
 
   // Cleanup event listeners whenever we start loading a new source
-  player.one('loadstart', () => {
+  player.dash.mediaPlayer.on(dashjs.MediaPlayer.events.STREAM_TEARDOWN_COMPLETE, () => {
     player.textTracks().off('change', updateActiveDashTextTrack);
   });
 
@@ -153,15 +157,11 @@ export default function setupTextTracks(player, tech, options) {
   // Attach dash text tracks whenever we dash emits `TEXT_TRACKS_ADDED`.
   mediaPlayer.on(dashjs.MediaPlayer.events.TEXT_TRACKS_ADDED, handleTextTracksAdded);
 
-  function cleanup() {
-    mediaPlayer.off(dashjs.MediaPlayer.events.TEXT_TRACKS_ADDED, handleTextTracksAdded);
-
-    player.one('loadstart', clearDashTracks);
-  }
-
   // When the player can play, remove the initialization events. We might not have received
   // TEXT_TRACKS_ADDED` so we have to stop listening for it or we'll get errors when we load new
   // videos and are listening for the same event in multiple places, including cleaned up
   // mediaPlayers.
-  mediaPlayer.on(dashjs.MediaPlayer.events.CAN_PLAY, cleanup);
+  mediaPlayer.on(dashjs.MediaPlayer.events.CAN_PLAY, () => {
+    mediaPlayer.off(dashjs.MediaPlayer.events.TEXT_TRACKS_ADDED, handleTextTracksAdded);
+  });
 }